Modified generate script to be better and also started research paper

+++ program/generate.sh
@@ -1,6 +1,9 @@
 #!/bin/bash
 
-rm config
+if [ -e config ];
+then
+	rm config
+fi
 numLines=$1
 for i in `seq 1 $numLines`;
 do
@@ -8,6 +11,5 @@ do
 	y=$((1 + RANDOM % 9))
 	name=$(pwgen 10 1)
 	line="$x $y $name"
-	echo $line
-	echo $line >> config
+	echo "$line" | tee -a config
 done
